refactor: Move sink frontend code to a seperate file
